return back utils to dq_actors
