+++ ydb/core/fq/libs/checkpoint_storage/events/ya.make
@@ -8,7 +8,6 @@ PEERDIR(
     ydb/library/actors/core
     ydb/library/actors/interconnect
     ydb/core/fq/libs/checkpointing_common
-    # ydb/core/fq/libs/events
     ydb/core/fq/libs/checkpoint_storage/proto
     yql/essentials/public/issue
 )
