+++ ydb/core/fq/libs/checkpointing/ut/checkpoint_coordinator_ut.cpp
@@ -9,7 +9,7 @@
 
 #include <google/protobuf/util/message_differencer.h>
 
-#include <ydb/library/yql/dq/actors/common/utils.h>
+#include <ydb/library/yql/dq/actors/compute/dq_compute_actor_checkpoints.h>
 #include <ydb/library/yql/providers/dq/api/protos/dqs.pb.h>
 #include <ydb/library/yql/providers/dq/common/yql_dq_settings.h>
