+++ ydb/library/yql/providers/dq/actors/task_controller_impl.h
@@ -16,7 +16,7 @@
 #include <yql/essentials/utils/log/log.h>
 
 #include <ydb/public/lib/yson_value/ydb_yson_value.h>
-#include <ydb/library/yql/dq/actors/common/utils.h>
+#include <ydb/library/yql/dq/actors/compute/dq_compute_actor_checkpoints.h>
 #include <ydb/library/yql/dq/actors/compute/dq_compute_actor.h>
 
 #include <ydb/library/actors/core/actorsystem.h>
